Section 11 - até aula 149

+++ src/Section11/code/resources/item.py
@@ -1,5 +1,6 @@
+from flask_jwt_extended.utils import get_jwt_identity
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from models.item import ItemModel
 
 class Item(Resource):
@@ -22,6 +23,7 @@ class Item(Resource):
             return item.json()
         return {'message': "Item not found"}, 404
 
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -39,6 +41,10 @@ class Item(Resource):
 
     @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -62,6 +68,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
-        # return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = list(map(lambda item: item.json(), ItemModel.query.all()))
+        if user_id:
+            return {'items': items }, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in.'
+        }, 200
